Added a new class with a special algorithm to get primes.
It's called Sieve of Eratothenes or something. Wikipedia it.

Right now the program can compile but fails. The bug is in line 36 of SieveGenerator.cs. The problem is that I tried to assign the index of a list a number, but supposedly I'm supposed to .Add(number) .

I guess the syntax is tricky because the list is a class list, which has two properties. How do you add something to just one of the properties?